Complete CYOA (Random Forest) mini-project
Chose to investigate Random Forest, which operates by averaging an ensemble
of decision tree classifiers created using random parameters and trained with
a sample drawn w/replacement from the training data. Created and trained an
instance of sklearn.ensemble.RandomForestClassifier, and played with the
constructor parameters to try to get a better accuracy than 93.6%.

I was able to meet the instructor's best accuracy, but could not exceed it.
Part of the difficulty in measure the accuracy of this classifier is that,
because the algorithm uses randomness to reduce bias/variance, the accuracy
is likewise random, and varies each runtime. I ran each set of chosen
parameters multiple times, but did not beat the 93.6% threshold.

It was comforting to discover that the Random Forest, despite creating multiple
decision tree estimators, was actually much faster to train than the single
decision tree from lesson three.  It was also more accurate.